special format for parameter redaction
If remaining parameters are redacted, it is possible to define a parameter name instead of a regex pattern to match parameter values. If a pattern is added in the format `%%parametername%%` and a parameter called `parametername` is present in the URL, it will remain but the value is replaced with the defined replacement text.